+++ rest/v1/redeem.ts
@@ -1,0 +1,6 @@
+import type { APIPayload, APIRedeemResponse } from "../../v1";
+
+/**
+ * @see https://docs.vertracloud.app/api-reference/endpoint/redeem
+ */
+export type RESTPostAPIRedeemResponse = APIPayload<APIRedeemResponse>;